# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bus Ticket Booking System API
  description: |
    Complete API documentation for the Bus Ticket Booking System.
    
    ## Features
    - Customer authentication and management
    - Bus schedule search and management
    - Real-time seat locking and booking
    - Payment integration with PayOS
    - Review and rating system
    - AI-powered chatbot
    - Admin dashboard analytics
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@busticketbooking.com
  license:
    name: ISC
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.busticketbooking.com
    description: Production server
...
paths:
...
  /api/bookings:
    post:
      tags:
        - Bookings
      summary: Create booking (authenticated user)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
  /api/bookings/guest:
    post:
      tags:
        - Bookings
      summary: Create guest booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CreateBookingRequest'
                - type: object
                  required:
                    - guestEmail
                    - guestPhone
                    - guestName
                  properties:
                    guestEmail:
                      type: string
                      format: email
                    guestPhone:
                      type: string
                    guestName:
                      type: string
      responses:
        '201':
          description: Guest booking created
  /api/bookings/my-bookings:
    get:
      tags:
        - Bookings
      summary: Get user's bookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed, expired]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Bookings retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                  total:
                    type: integer
                  page:
                    type: integer
                  totalPages:
                    type: integer
  /api/bookings/{id}/confirm:
    post:
      tags:
        - Bookings
      summary: Confirm booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking confirmed
  /api/bookings/{id}/cancel:
    post:
      tags:
        - Bookings
      summary: Cancel booking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Booking cancelled
          
  # Seat Management
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateBookingRequest:
      type: object
      required:
        - busScheduleId
        - seatNumbers
        - passengers
      properties:
        busScheduleId:
          type: integer
        seatNumbers:
          type: array
          items:
            type: integer
        passengers:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              age:
                type: integer
              gender:
                type: string
              seatNumber:
                type: integer
        pickupPoint:
          type: string
        dropoffPoint:
          type: string
    Booking:
      type: object
      properties:
        id:
          type: integer
        bookingReference:
          type: string
        customerId:
          type: integer
        busScheduleId:
          type: integer
        seatNumbers:
          type: array
          items:
            type: integer
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed, expired]
        journeyDate:
          type: string
          format: date
        payment_totalPay:
          type: number
        passengers:
          type: array
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    BookingResponse:
      type: object
      properties:
        msg:
          type: string
        booking:
          $ref: '#/components/schemas/Booking'
...